ZoomScreenshot/requirements.txt
pyautogui==0.9.54
keyboard==0.13.5
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
